[Refactor] Extract config to component and optimize expression logic
